requirements.txt
aioredis==1.3.1
alembic==1.7.6
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
click==8.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0.2
django-autoslug==1.9.8
django-cors-headers==3.11.0
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.2.0
email-validator==1.1.3
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==3.4.5
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
greenlet==1.1.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
incremental==21.3.0
itsdangerous==2.1.0
Jinja2==3.0.3
ldap3==2.9.1
Mako==1.1.6
MarkupSafe==2.1.0
msgpack==1.0.3
paramiko==2.9.2
passlib==1.7.2
pgadmin4==4.30
phonenumbers==8.12.44
Pillow==9.0.1
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2020.5
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.31
sqlparse==0.4.2
sshtunnel==0.4.0
Twisted==22.1.0
txaio==22.2.1
typing-extensions==4.1.1
Werkzeug==2.0.3
WTForms==2.2.1
zipp==3.7.0
zope.interface==5.4.0
